upload 2.2,4d527ba5fcc7

+++ interdrv/v2/vpss/common/vpss_rgn_ctrl.h
@@ -1,8 +1,8 @@
 #ifndef _VPSS_RGN_CTRL_H_
 #define _VPSS_RGN_CTRL_H_
 
-s32 vpss_get_rgn_hdls(vpss_grp vpss_grp, vpss_chn vpss_chn, u8 layer,
-	rgn_type_e type, rgn_handle hdls[]);
+s32 vpss_get_rgn_hdls(vpss_grp vpss_grp, vpss_chn vpss_chn, u32 layer,
+			rgn_type_e type, rgn_handle hdls[]);
 s32 vpss_set_rgn_hdls(vpss_grp vpss_grp, vpss_chn vpss_chn, u32 layer,
 		rgn_type_e type, rgn_handle hdls[]);
 s32 vpss_set_rgn_cfg(vpss_grp vpss_grp, vpss_chn vpss_chn, u32 layer, struct rgn_cfg *cfg);
@@ -11,5 +11,4 @@ s32 vpss_set_rgn_mosaic_cfg(vpss_grp vpss_grp, vpss_chn vpss_chn, struct rgn_mos
 s32 vpss_get_rgn_ow_addr(vpss_grp vpss_grp, vpss_chn vpss_chn, u32 layer,
 	rgn_handle handle, u64 *addr);
 s32 vpss_set_rgn_lut_cfg(vpss_grp vpss_grp, vpss_chn vpss_chn, struct rgn_lut_cfg *cfg);
-
 #endif /* _VPSS_RGN_CTRL_H_ */
